refactor: separate out `StdoutReader` to own file
